Update ClassSessionAdd.java and AddClassSession.jsp
Customize the select tag when updating/save classrooms. Each classroom can only select 1 classroom. Selected classrooms will hide that room in the list.

Add a check condition, if there are already students participating in that class, the status cannot be changed, but the room and session can still be updated.